+++ customer-work-example/src/main/java/com/richard/fyoung/customerworkapp/controller/CustomerServiceController.java
@@ -1,4 +1,4 @@
-package com.richard.fyoung.customerwork.controller;
+package com.richard.fyoung.customerworkapp.controller;
 
 import com.richard.fyoung.customerwork.agent.AguiService;
 import com.richard.fyoung.customerwork.agent.MultiAgentOrchestrator;